% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliplot.R
\name{add_stat_annotations}
\alias{add_stat_annotations}
\title{Add statistical test annotations}
\usage{
add_stat_annotations(p, settings, data)
}
\description{
Add statistical test annotations
}
\keyword{internal}
