% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cli_palette_list}
\alias{cli_palette_list}
\title{List Available Cliomicplot Palettes}
\usage{
cli_palette_list()
}
\value{
Character vector of palette names.
}
\description{
Returns the names of all built-in cliomicplot color palettes.
}
