% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cli_palette_show}
\alias{cli_palette_show}
\title{Show a Cliomicplot Palette}
\usage{
cli_palette_show(name = "jco")
}
\arguments{
\item{name}{Character. Palette name from \code{\link{cli_palette_list}}.}
}
\value{
Invisibly returns the plotted palette from \code{scales::show_col()}.
}
\description{
Displays a named palette as labeled color swatches.
}
