% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliplot_add.R
\name{cliplot_add}
\alias{cliplot_add}
\title{Add Layers to an Existing Cliomicplot}
\usage{
cliplot_add(type = "points", ...)
}
\arguments{
\item{type}{Plot type for the added layer.}

\item{...}{Additional arguments passed to the type's draw function.}
}
\description{
Adds additional layers (points, lines, text, etc.) to the last cliomicplot
created. Similar to tinyplot's `tinyplot_add()`.
}
\details{
Currently a placeholder. Full layer addition requires tracking the last
ggplot object, which will be implemented in a future version.
}
