% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clitheme.R
\name{clitheme_list}
\alias{clitheme_list}
\title{List Available Themes}
\usage{
clitheme_list()
}
\value{
Character vector of theme names.
}
\description{
Returns a character vector of all available and registered themes.
}
