% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_theme.R
\name{clitheme_md_reset}
\alias{clitheme_md_reset}
\title{Disable Global Markdown Rendering}
\usage{
clitheme_md_reset()
}
\description{
Resets the active ggplot2 theme to the base cliomicplot theme
without markdown support.
}
