% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clitheme.R
\name{clitheme_unregister}
\alias{clitheme_unregister}
\title{Unregister a Custom Theme}
\usage{
clitheme_unregister(name)
}
\arguments{
\item{name}{Character string. Name of the theme to unregister.}
}
\description{
Remove a previously registered custom theme.
}
