% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_data}
\alias{extract_data}
\title{Extract columns safely from a data frame}
\usage{
extract_data(data, x_var, y_var = NULL, by_var = NULL)
}
\description{
Extract columns safely from a data frame
}
\keyword{internal}
