% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_theme_overrides}
\alias{parse_theme_overrides}
\title{Parse theme overrides from ... arguments}
\usage{
parse_theme_overrides(...)
}
\description{
Parse theme overrides from ... arguments
}
\keyword{internal}
