% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clitheme.R
\name{resolve_theme}
\alias{resolve_theme}
\title{Resolve a theme: returns a ggplot2 theme object from a name or object}
\usage{
resolve_theme(theme)
}
\description{
Resolve a theme: returns a ggplot2 theme object from a name or object
}
\keyword{internal}
