% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliplot.R
\name{resolve_type}
\alias{resolve_type}
\title{Resolve a type string or cliplot_type object to a type definition}
\usage{
resolve_type(type)
}
\description{
Resolve a type string or cliplot_type object to a type definition
}
\keyword{internal}
