% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliomicplot-package.R
\name{theme_cli_base}
\alias{theme_cli_base}
\title{Base cliomicplot ggplot2 theme}
\usage{
theme_cli_base(base_size = 11, base_family = "")
}
\description{
Base cliomicplot ggplot2 theme
}
\keyword{internal}
