% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_system.R
\name{type_barplot}
\alias{type_barplot}
\title{Bar plot type}
\usage{
type_barplot(stat = "count", position = "stack", alpha = 0.9)
}
\arguments{
\item{stat}{Statistic to compute ("count", "identity")}

\item{position}{Bar position ("stack", "dodge", "fill")}

\item{alpha}{Fill transparency}
}
\description{
Bar plot type
}
