% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_system.R
\name{type_density}
\alias{type_density}
\title{Density plot type}
\usage{
type_density(alpha = 0.28, linewidth = 1)
}
\arguments{
\item{alpha}{Fill transparency}

\item{linewidth}{Line width}
}
\description{
Density plot type
}
