% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_system.R
\name{type_errorbar}
\alias{type_errorbar}
\title{Error bar type (requires ymin and ymax)}
\usage{
type_errorbar(width = 0.2, linewidth = 0.6)
}
\arguments{
\item{width}{Error bar width}

\item{linewidth}{Line width}
}
\description{
Error bar type (requires ymin and ymax)
}
