% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_system.R
\name{type_histogram}
\alias{type_histogram}
\title{Histogram type}
\usage{
type_histogram(bins = 30, alpha = 0.85)
}
\arguments{
\item{bins}{Number of bins}

\item{alpha}{Fill transparency}
}
\description{
Histogram type
}
