% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_system.R
\name{type_lines}
\alias{type_lines}
\title{Line plot type}
\usage{
type_lines(linewidth = 1, alpha = 0.95)
}
\arguments{
\item{linewidth}{Line width}

\item{alpha}{Line transparency}
}
\description{
Line plot type
}
