% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_system.R
\name{type_points}
\alias{type_points}
\title{Scatter plot type}
\usage{
type_points(alpha = NULL, size = NULL, stroke = 0.35)
}
\arguments{
\item{alpha}{Point transparency (0-1)}

\item{size}{Point size}

\item{stroke}{Point outline stroke width}
}
\description{
Scatter plot type
}
