% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_more.R
\name{type_qq}
\alias{type_qq}
\title{Q-Q Plot Type}
\usage{
type_qq(
  distribution = stats::qnorm,
  dparams = list(),
  point_size = 2,
  point_alpha = 0.6,
  show_line = TRUE,
  line_color = "#E64B35"
)
}
\arguments{
\item{distribution}{Target distribution function (e.g., stats::qnorm)}

\item{dparams}{List of distribution parameters}

\item{point_size}{Point size}

\item{point_alpha}{Point transparency}

\item{show_line}{Add reference line (default TRUE)}

\item{line_color}{Color for reference line}
}
\value{
A cliplot_type object
}
\description{
Creates quantile-quantile plots to assess normality or
  compare two distributions.
}
