% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_more.R
\name{type_spineplot}
\alias{type_spineplot}
\title{Spine Plot / Spinogram Type}
\usage{
type_spineplot(weights = NULL)
}
\arguments{
\item{weights}{Optional vector of weights}
}
\value{
A cliplot_type object
}
\description{
Creates spine plots or spinograms for visualizing the
  relationship between categorical variables. Width of bars represents
  marginal distribution; height represents conditional distribution.
}
